[Fix] flashメッセージの不具合を修正

+++ app/controllers/favorites_controller.rb
@@ -7,9 +7,14 @@ class FavoritesController < ApplicationController
     # ログインユーザに紐づくfavoriteテーブルのレコードを作成、対象となるガジェット記事は
     # 1行上で取得した情報に基づく。
     favorite = current_user.favorites.new(gadget_id: @gadget.id)
-    favorite.save
-    @gadget.create_notification_favorite!(current_user)
-
+    unless current_user.id == @gadget.user_id
+      favorite.save
+      @gadget.create_notification_favorite!(current_user)
+    else
+      # 非同期処理エラー時の動作を設定したjsファイルを指定
+      flash.now[:alert] = "自身の投稿をお気に入り登録することはできません。"
+      render 'favorites/error'
+    end
   end
 
   def destroy
@@ -18,7 +23,14 @@ class FavoritesController < ApplicationController
     # 1行上取得した情報をキーとして、ログインユーザに紐づくfavoriteテーブルから対象となる
     # レコードを検索する
     favorite = current_user.favorites.find_by(gadget_id: @gadget.id)
-    favorite.destroy
+
+    unless current_user.id == @gadget.user_id
+      favorite.destroy
+    else
+      # 非同期処理エラー時の動作を設定したjsファイルを指定
+      flash.now[:alert] = "不正な操作です。"
+      render 'favorites/error'
+    end
 
   end
 end
